backend-host/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.26.0
redis>=5.0.1
python-multipart>=0.0.6
groq>=0.4.0
# RAG dependencies
chromadb>=0.4.22
sentence-transformers>=2.3.1
langchain>=0.1.0
langchain-community>=0.0.13
tiktoken>=0.5.2
pypdf>=3.17.0
# Fix for Keras 3 compatibility
tf-keras>=2.15.0
# Fix for deprecated pynvml warning
nvidia-ml-py>=11.495.46

